Added tests to youtrackconfiguration.
