+++ sentry_youtrack/plugin.py
@@ -7,14 +7,14 @@ from django.http import HttpResponse
 from django.utils.translation import ugettext_lazy as _
 from sentry.models import GroupMeta
 from sentry.plugins.bases.issue import IssuePlugin
-from sentry.exceptions import PluginError
+
 
 from . import VERSION
 from .forms import (NewIssueForm, AssignIssueForm, DefaultFieldForm,
                     YouTrackProjectForm, VERIFY_SSL_CERTIFICATE)
 from .utils import cache_this, get_int
 from .youtrack import YouTrackClient
-from requests.exceptions import ConnectionError, HTTPError, SSLError
+from sentry_youtrack.configuration import YouTrackConfiguration
 
 
 class YouTrackPlugin(IssuePlugin):
@@ -189,150 +189,3 @@ class YouTrackPlugin(IssuePlugin):
             # raise PluginError(message)
             pass
         return config
-
-class YouTrackConfiguration(object):
-
-    error_message = {
-        'client': _("Unable to connect to YouTrack."),
-        'project_unknown': _('Unable to fetch project'),
-        'project_not_found': _('Project not found: %s'),
-        'invalid_ssl': _("SSL certificate  verification failed."),
-        'invalid_password': _('Invalid username or password.'),
-        'invalid_project': _('Invalid project: \'%s\''),
-        'missing_fields': _('Missing required fields.'),
-        'perms': _("User doesn't have Low-level Administration permissions."),
-        'required': _("This field is required.")}
-
-    def __init__(self, initial):
-        self.config = self.build_default_fields()
-        self.set_initial(initial)
-
-    def set_initial(self, initial):
-        self.client_errors = {}
-        if self.has_client_fields(initial):
-            client = self.get_youtrack_client(initial)
-            if client:
-                if initial.get('project'):
-                    choices = self.get_ignore_field_choices(
-                        client, initial.get('project'))
-                    self.config.append({
-                        'name':'ignore_fields',
-                        'label':'Ignore Fields',
-                        'type':'select',
-                        'choices':choices,
-                        'required':False,
-                        'help': 'These fields will not appear on the form.',})
-
-                choices = self.get_project_field_choices(client, initial.get('project'))
-                self.config.append({
-                    'name':'project',
-                    'label':'Linked Project',
-                    'type':'select',
-                    'choices': choices,
-                    'required':True,})
-                
-                self.config.append({'name':'default_tags',
-                'label':'Default Tags',
-                'type':'text',
-                'required':False,
-                'placeholder': 'e.g. sentry',
-                'help': 'Comma-separated list of tags.',})
-
-                if not initial.get('project'):
-                    self.second_step_msg = _(
-                        "Your credentials are valid but plugin is NOT active "
-                        "yet. Please fill in remaining required fields.")
-
-    def has_client_fields(self, initial):
-        return initial.get('password') and initial.get('username') and initial.get('url')
-                
-    def build_default_fields(self):
-        url = {'name':'url',
-                'label':'YouTrack Instance URL',
-                'type':'text',
-                'required':True,
-                'placeholder': 'e.g. "https://yoursitename.myjetbrains.com/youtrack/"',}
-        username = {'name':'username',
-                'label':'Username',
-                'type':'text',
-                'required':True,
-                'help': 'User should have admin rights.',}
-        password = {'name':'password',
-                'label':'Password',
-                'type':'secret',
-                'required':False,
-                'help': 'Only enter a password if you want to change it.',}
-       
-        return [url, username, password]
-        
-
-    def get_youtrack_client(self, data, additional_params=None):
-        yt_settings = {
-            'url': data.get('url'),
-            'username': data.get('username'),
-            'password': data.get('password'),
-            'verify_ssl_certificate': VERIFY_SSL_CERTIFICATE}
-        if additional_params:
-            yt_settings.update(additional_params)
-
-        client = None
-        try:
-            client = YouTrackClient(**yt_settings)
-        except (HTTPError, ConnectionError) as e:
-            if e.response is not None and e.response.status_code == 403:
-                self.client_errors['username'] = self.error_message[
-                    'invalid_password']
-            else:
-                self.client_errors['url'] = self.error_message['client']
-        except (SSLError, TypeError) as e:
-            self.client_errors['url'] = self.error_message['invalid_ssl']
-        if client:
-            try:
-                client.get_user(yt_settings.get('username'))
-            except HTTPError as e:
-                if e.response.status_code == 403:
-                    self.client_errors['username'] = self.error_message['perms']
-                    client = None
-        return client
-
-    def get_ignore_field_choices(self, client, project):
-        try:
-            fields = list(client.get_project_fields_list(project))
-        except HTTPError:
-            self.client_errors['project'] = self.error_message[
-                'invalid_project'] % (project,)
-        else:
-            names = [field['name'] for field in fields]
-            return zip(names, names)
-        return []
-
-    def get_project_field_choices(self, client, project):
-        choices = [(' ', u"- Choose project -")]
-        try:
-            projects = list(client.get_projects())
-        except HTTPError:
-            self.client_errors['project'] = self.error_message[
-                'invalid_project'] % (project, )
-        else:
-            for project in projects:
-                display = "%s (%s)" % (project['name'], project['id'])
-                choices.append((project['id'], display))
-        return choices
-    
-    def get_project_fields_list(self, client, project_id):
-        try:
-            return list(client.get_project_fields_list(project_id))
-        except (HTTPError, ConnectionError) as e:
-            if e.response is not None and e.response.status_code == 404:
-                self.client_errors['project'] = self.error_message['project_not_found'] % project_id
-            else:
-                self.client_errors['project'] = self.error_message['project_unknown']
-
-    def get_projects(self, client):
-        try:
-            return list(client.get_projects())
-        except (HTTPError, ConnectionError) as e:
-            if e.response is not None and e.response.status_code == 404:
-                self.client_errors['project'] = self.error_message['project_not_found'] % project_id
-            else:
-                self.client_errors['project'] = self.error_message['project_unknown']
