+++ tests/cassettes/yt_config.yaml
@@ -1,0 +1,59 @@
+interactions:
+- request:
+    body: login=root&password=admin
+    headers:
+      Content-Type: [application/x-www-form-urlencoded]
+      User-Agent: [sentry-youtrack/0.2.4]
+    method: POST
+    uri: https://youtrack.myjetbrains.com/rest/user/login
+  response:
+    body: {string: !!python/unicode '<login>ok</login>'}
+    headers:
+      access-control-expose-headers: [Location]
+      cache-control: ['no-cache, no-store, no-transform, must-revalidate']
+      content-type: [application/xml; charset=UTF-8]
+      expires: ['Thu, 01 Jan 1970 00:00:00 GMT']
+      server: [Jetty(8.y.z-SNAPSHOT)]
+      set-cookie: [YTSESSIONID=p66uquc6xxzh1err5odptbsr8;Path=/, 'jetbrains.charisma.main.security.PRINCIPAL=abcd1234;Path=/;']
+      vary: [Accept-Encoding]
+    status: {code: 200, message: OK}
+- request:
+    body: null
+    headers:
+      Cookie: [jetbrains.charisma.main.security.PRINCIPAL=abcd1234]
+      User-Agent: [sentry-youtrack/0.2.4]
+    method: GET
+    uri: https://youtrack.myjetbrains.com/rest/admin/user/root
+  response:
+    body: {string: !!python/unicode '<?xml version="1.0" encoding="UTF-8" standalone="yes"?> <user login="root" fullName="root" groupsUrl="https://youtrack.myjetbrains.com/rest/admin/user/root/group/" rolesUrl="https://youtrack.myjetbrains.com/rest/admin/user/root/role/"/>'}
+    headers:
+      access-control-expose-headers: [Location]
+      cache-control: ['no-cache, no-store, no-transform, must-revalidate']
+      content-length: ['180']
+      content-type: [application/xml; charset=UTF-8]
+      expires: ['Thu, 01 Jan 1970 00:00:00 GMT']
+      server: [Jetty(8.y.z-SNAPSHOT)]
+      set-cookie: [YTSESSIONID=1nh11tmjdq4ys1cra3scg01kmv;Path=/]
+      vary: [Accept-Encoding]
+    status: {code: 200, message: OK}
+- request:
+    body: null
+    headers:
+      Cookie: [jetbrains.charisma.main.security.PRINCIPAL=abcd1234]
+      User-Agent: [sentry-youtrack/0.2.4]
+    method: GET
+    uri: https://youtrack.myjetbrains.com/rest/project/all
+  response:
+    body: {string: !!python/unicode '<?xml version="1.0" encoding="UTF-8" standalone="yes"?><projects><project
+        name="My project" shortName="myproject"/><project name="Test project" shortName="testproject"/></projects>'}
+    headers:
+      access-control-expose-headers: [Location]
+      cache-control: ['no-cache, no-store, no-transform, must-revalidate']
+      content-length: ['180']
+      content-type: [application/xml; charset=UTF-8]
+      expires: ['Thu, 01 Jan 1970 00:00:00 GMT']
+      server: [Jetty(8.y.z-SNAPSHOT)]
+      set-cookie: [YTSESSIONID=1nh11tmjdq4ys1cra3scg01kmv;Path=/]
+      vary: [Accept-Encoding]
+    status: {code: 200, message: OK}
+version: 1
